refactor(api): decompose TMDB god-class into tmdb/ submodule
Split the 1123-line TmdbApi into a thin facade over a transport client and focused sub-clients (movies, TV, genres, reviews, cross-type find), matching the existing igdb/ / anilist/ / openlibrary/ layout. Behaviour and the public API are unchanged: the facade re-exports the same types and delegates every method, so callers and mocks are untouched.
